+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+- `groups-service` `ensureUniqueGroupName` & `doesGroupExist` methods
+
 ## 0.23.1
 ### Fixed
 - bad params in `project` function